Change IFeeModel.GetOrderFee to return OrderFee type
This change accomplishes a few goals. First, it provides units for computed
order fees. As of this commit all fees are still assumed to be in units of
the account currency which is enforced via usage of the IdentityCurrencyConverter
in the shim extension method. Second, the definition of IFeeModel, and the
interaction with OrderFeeContext, is such that we can modify the types and
behaviors without fear of further breaking changes. The methods on OrderFeeContext
that produce OrderFee objects act as a natural shim and convenience methods at
the same time. These factory methods can be maintained even if later we need
to change the various semantics and/or dependencies around fee computation