Refactor IFeeModel.GetOrderFee
In preparation for sweeping changes that will require adding a currency converter
parameter to this method and augmenting the return value to be CashAmount. This
change ensures the future change won't produce a compile-time breaking change. An
extension method was also added as a shim to keep any existing code functional